pyproject.toml
[project]
name = "gmail2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
